Fix error handler bug.
Previously it wasn't being called deep enough in the call stack to
provide an error response, so when errors occurred, the server would
just keep the socket open forever and never write a response.

Additionally, if the router doesn't return anything, we should at
least return a 404 rather than leaving the client hanging. If you want
to leave the client hanging intentionally, you can use false for that.